backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
python-multipart
Pillow
sqlalchemy
openai
